+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [v0.6.0]
+### Added
+- Assembly step to de-novo consensus mode to handle longer amplicons. README was updated accordingly.
+
 ### Removed
 - Default local executor CPU and RAM limits